refactor: move changeCowAutomaticHugState to its own file
