+++ src/game-logic/reducers/changeCowAutomaticHugState.test.js
@@ -1,0 +1,70 @@
+import { huggingMachine } from '../../data/items'
+import { generateCow } from '../../utils'
+
+import { changeCowAutomaticHugState } from './changeCowAutomaticHugState'
+
+describe('changeCowAutomaticHugState', () => {
+  describe('setting isUsingHuggingMachine to true', () => {
+    test('sets isUsingHuggingMachine to true', () => {
+      const cow = generateCow()
+      const inputState = {
+        cowInventory: [cow],
+        inventory: [{ id: huggingMachine.id, quantity: 1 }],
+        inventoryLimit: -1,
+      }
+      const {
+        cowInventory: [{ isUsingHuggingMachine }],
+        inventory,
+      } = changeCowAutomaticHugState(inputState, cow, true)
+
+      expect(isUsingHuggingMachine).toEqual(true)
+      expect(inventory).toEqual([])
+    })
+
+    describe('there are no hugging machines in inventory', () => {
+      test('no-ops', () => {
+        const cow = generateCow()
+        const inputState = {
+          cowInventory: [cow],
+          inventory: [],
+          inventoryLimit: -1,
+        }
+        const state = changeCowAutomaticHugState(inputState, cow, true)
+
+        expect(state).toBe(inputState)
+      })
+    })
+
+    describe('isUsingHuggingMachine is already true', () => {
+      test('no-ops', () => {
+        const cow = generateCow({ isUsingHuggingMachine: true })
+        const inputState = {
+          cowInventory: [cow],
+          inventory: [{ id: huggingMachine.id, quantity: 1 }],
+          inventoryLimit: -1,
+        }
+        const state = changeCowAutomaticHugState(inputState, cow, true)
+
+        expect(state).toBe(inputState)
+      })
+    })
+  })
+
+  describe('setting isUsingHuggingMachine to false', () => {
+    test('sets isUsingHuggingMachine to false', () => {
+      const cow = generateCow({ isUsingHuggingMachine: true })
+      const inputState = {
+        cowInventory: [cow],
+        inventory: [],
+        inventoryLimit: -1,
+      }
+      const {
+        cowInventory: [{ isUsingHuggingMachine }],
+        inventory,
+      } = changeCowAutomaticHugState(inputState, cow, false)
+
+      expect(isUsingHuggingMachine).toEqual(false)
+      expect(inventory).toEqual([{ id: huggingMachine.id, quantity: 1 }])
+    })
+  })
+})
