+++ src/game-logic/reducers/index.test.js
@@ -14,7 +14,6 @@ import {
   PURCHASEABLE_COW_PENS,
   STORAGE_EXPANSION_AMOUNT,
 } from '../../constants'
-import { huggingMachine } from '../../data/items'
 import {
   generateCow,
   getCostOfNextStorageExpansion,
@@ -43,72 +42,6 @@ describe('selectCow', () => {
   test('updates selectedCowId', () => {
     const { selectedCowId } = fn.selectCow({}, { id: 'abc' })
     expect(selectedCowId).toEqual('abc')
-  })
-})
-
-describe('changeCowAutomaticHugState', () => {
-  describe('setting isUsingHuggingMachine to true', () => {
-    test('sets isUsingHuggingMachine to true', () => {
-      const cow = generateCow()
-      const inputState = {
-        cowInventory: [cow],
-        inventory: [{ id: huggingMachine.id, quantity: 1 }],
-        inventoryLimit: -1,
-      }
-      const {
-        cowInventory: [{ isUsingHuggingMachine }],
-        inventory,
-      } = fn.changeCowAutomaticHugState(inputState, cow, true)
-
-      expect(isUsingHuggingMachine).toEqual(true)
-      expect(inventory).toEqual([])
-    })
-
-    describe('there are no hugging machines in inventory', () => {
-      test('no-ops', () => {
-        const cow = generateCow()
-        const inputState = {
-          cowInventory: [cow],
-          inventory: [],
-          inventoryLimit: -1,
-        }
-        const state = fn.changeCowAutomaticHugState(inputState, cow, true)
-
-        expect(state).toBe(inputState)
-      })
-    })
-
-    describe('isUsingHuggingMachine is already true', () => {
-      test('no-ops', () => {
-        const cow = generateCow({ isUsingHuggingMachine: true })
-        const inputState = {
-          cowInventory: [cow],
-          inventory: [{ id: huggingMachine.id, quantity: 1 }],
-          inventoryLimit: -1,
-        }
-        const state = fn.changeCowAutomaticHugState(inputState, cow, true)
-
-        expect(state).toBe(inputState)
-      })
-    })
-  })
-
-  describe('setting isUsingHuggingMachine to false', () => {
-    test('sets isUsingHuggingMachine to false', () => {
-      const cow = generateCow({ isUsingHuggingMachine: true })
-      const inputState = {
-        cowInventory: [cow],
-        inventory: [],
-        inventoryLimit: -1,
-      }
-      const {
-        cowInventory: [{ isUsingHuggingMachine }],
-        inventory,
-      } = fn.changeCowAutomaticHugState(inputState, cow, false)
-
-      expect(isUsingHuggingMachine).toEqual(false)
-      expect(inventory).toEqual([{ id: huggingMachine.id, quantity: 1 }])
-    })
   })
 })
